Make 'steps' and 'environment' importable.
